% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class.R
\name{ds.omop.helper}
\alias{ds.omop.helper}
\title{Factory function for OMOP CDM Helper}
\usage{
ds.omop.helper(
  connections,
  resource,
  symbol,
  personColumns = NULL,
  personFilter = NULL
)
}
\arguments{
\item{connections}{Connection object to the DataSHIELD server.}

\item{resource}{Either an identifier or a named list of identifiers for the specific resource(s) within the DataSHIELD server(s).
If a named list, the name of each resource identifier should correspond to the server name in the connections.}

\item{symbol}{A character string representing the symbol to which the resource will be assigned.}

\item{personColumns}{Optional; a vector of column names to filter from the person table.}

\item{personFilter}{Optional; a character string representing another object in the DataSHIELD environment whose person IDs will be
used to filter the processed data (this speeds up the processing of the data if not all the person IDs are needed).}
}
\value{
A new instance of the OMOPCDMHelper class.
}
\description{
This function creates a new instance of the OMOPCDMHelper class, allowing
for interaction with an OMOP CDM database through the DataSHIELD environment.
}
