% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_auto.R
\name{generateTableColumns}
\alias{generateTableColumns}
\title{Generate table column names with prefixes}
\usage{
generateTableColumns(tableName, columnNames)
}
\arguments{
\item{tableName}{A character string specifying the name of the table.}

\item{columnNames}{An optional vector of column names to include in the operation.}
}
\value{
A vector of column names, optionally prefixed with the modified table name. If no column names are provided,
the function returns NULL.
}
\description{
This function generates a list of column names for a specified table, optionally prefixing them with the table name.
It is designed to handle cases where no column names are provided, in which case it simply returns NULL.
The function also removes the "_occurrence" and "_exposure" suffixes from the table name before prefixing, ensuring a cleaner column name output.
}
