% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_get.R
\name{OMOPCDMHelper-get}
\alias{OMOPCDMHelper-get}
\title{Get Table from OMOP CDM Database}
\arguments{
\item{table}{A string specifying the name of the table to be retrieved.}

\item{symbol}{An optional string specifying the symbol for the table assignment in the DataSHIELD environment.}

\item{columnFilter}{An optional string vector specifying column names to filter (select) in the table.}

\item{conceptFilter}{An optional numeric vector specifying concept IDs to filter the table by.}

\item{personFilter}{An optional string specifying the symbol in the environment of a table from which to obtain person IDs.}

\item{mergeColumn}{An optional string specifying the column name for merging operations with other tables.
Defaults to "person_id" if not specified.}

\item{dropNA}{An optional boolean indicating whether to drop empty columns. Defaults to FALSE.}

\item{...}{Additional parameters to be passed to the underlying get method.}
}
\value{
Assigns the specified table to the DataSHIELD environment, optionally filtered and merged.
}
\description{
This method retrieves a specified table by name from an OMOP CDM database and assigns it to the DataSHIELD environment.
It allows the user to specify filters for concepts, columns, and persons, as well as options for merging and dropping empty columns.
}
