% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_append.R
\name{OMOPCDMHelper-append}
\alias{OMOPCDMHelper-append}
\title{Append a table to the base table}
\arguments{
\item{table}{A character string specifying the name of the table to append data to.}

\item{columns}{Optional; a vector of column names to include in the operation.}

\item{concepts}{Optional; a list of concept IDs to filter the data by.}

\item{merge.x}{Optional; a character string specifying the column name in the base table to merge on.}

\item{merge.y}{Optional; a character string specifying the column name in the target table to merge on.}
}
\value{
The function does not return a value but modifies the database by appending data to the specified table.
}
\description{
This function appends a specified table from the OMOP CDM database to the base table, with options to filter
columns, apply concept filters, and specify merge conditions.
}
