% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_catalog.R
\name{OMOPCDMHelper-columns}
\alias{OMOPCDMHelper-columns}
\title{Retrieve Columns from OMOP CDM Database}
\arguments{
\item{tables}{Optional; a vector of table names to include in the operation. If NULL, all tables are considered.}
}
\value{
A list containing the names of all columns available in the specified tables of the OMOP CDM database.
}
\description{
This method fetches a comprehensive list of columns from specified or all tables within the OMOP CDM database.
across all connected servers. It leverages the underlying OMOPCDMDatabase class
to query the database and fetch the catalog of columns.
}
