% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_catalog.R
\name{OMOPCDMHelper-concepts}
\alias{OMOPCDMHelper-concepts}
\title{Retrieve Concepts Dictionary from OMOP CDM Database}
\arguments{
\item{tables}{Optional; a vector of table names to include in the operation. If NULL, all tables are considered.}

\item{max_length}{Optional; an integer specifying the maximum length of the concept names. If NULL, no truncation is performed.}
}
\value{
A list containing the concepts available in the specified tables of the OMOP CDM database.
}
\description{
This function fetches the dictionary of concepts from specified or all tables within the OMOP CDM database.
across all connected servers. It leverages the underlying OMOPCDMDatabase class
to query the database and fetch the catalog of concepts.
}
