% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_catalog.R
\name{OMOPCDMHelper-tables}
\alias{OMOPCDMHelper-tables}
\title{Retrieve Tables from OMOP CDM Database}
\value{
A list containing the names of all tables available in the OMOP CDM database.
}
\description{
This function fetches a list of all tables present in the OMOP CDM database.
across all connected servers. It leverages the underlying OMOPCDMDatabase class
to query the database and fetch the catalog of tables.
}
