% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class.R
\name{OMOPCDMHelper}
\alias{OMOPCDMHelper}
\title{OMOP CDM Helper}
\description{
This class facilitates the creation of tables from a connection to an OMOP CDM database.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{OMOPCDMDatabase}}{An object representing the connection to the OMOP CDM database.}

\item{\code{symbol}}{A symbol used to reference the retrieved data in the DataSHIELD environment.

Constructor for OMOP CDM Helper}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OMOPCDMHelper-new}{\code{OMOPCDMHelper$new()}}
\item \href{#method-OMOPCDMHelper-append}{\code{OMOPCDMHelper$append()}}
\item \href{#method-OMOPCDMHelper-tables}{\code{OMOPCDMHelper$tables()}}
\item \href{#method-OMOPCDMHelper-columns}{\code{OMOPCDMHelper$columns()}}
\item \href{#method-OMOPCDMHelper-concepts}{\code{OMOPCDMHelper$concepts()}}
\item \href{#method-OMOPCDMHelper-auto}{\code{OMOPCDMHelper$auto()}}
\item \href{#method-OMOPCDMHelper-get}{\code{OMOPCDMHelper$get()}}
\item \href{#method-OMOPCDMHelper-clone}{\code{OMOPCDMHelper$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMHelper-new"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMHelper-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMHelper$new(
  connections,
  resource,
  symbol,
  personColumns = NULL,
  personFilter = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connections}}{Connection object to the DataSHIELD server.}

\item{\code{resource}}{Either an identifier or a named list of identifiers for the specific resource(s) within the DataSHIELD server(s).
If a named list, the name of each resource identifier should correspond to the server name in the connections.}

\item{\code{symbol}}{A character string representing the symbol to which the resource will be assigned.}

\item{\code{personColumns}}{Optional; a vector of column names to filter from the person table.}

\item{\code{personFilter}}{Optional; a character string representing another object in the DataSHIELD environment whose person IDs will be
used to filter the processed data (this speeds up the processing of the data if not all the person IDs are needed).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class OMOPCDMHelper.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMHelper-append"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMHelper-append}{}}}
\subsection{Method \code{append()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMHelper$append(
  table,
  columns = NULL,
  concepts = NULL,
  merge.x = NULL,
  merge.y = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMHelper-tables"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMHelper-tables}{}}}
\subsection{Method \code{tables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMHelper$tables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMHelper-columns"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMHelper-columns}{}}}
\subsection{Method \code{columns()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMHelper$columns(tables = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMHelper-concepts"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMHelper-concepts}{}}}
\subsection{Method \code{concepts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMHelper$concepts(tables = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMHelper-auto"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMHelper-auto}{}}}
\subsection{Method \code{auto()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMHelper$auto(
  tables = NULL,
  columns = NULL,
  concepts = NULL,
  silent = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMHelper-get"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMHelper-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMHelper$get(
  table,
  symbol = NULL,
  columnFilter = NULL,
  conceptFilter = NULL,
  personFilter = NULL,
  mergeColumn = NULL,
  dropNA = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMHelper-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMHelper-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMHelper$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
